RELATED: ONE-4414 Manual column resizing in KD
