+++ package.json
@@ -23,7 +23,7 @@
     "@babel/runtime": "7.4.4",
     "@gooddata/eslint-config": "^1.0.0",
     "@gooddata/frontend-npm-scripts": "1.2.0",
-    "@gooddata/mock-js": "^2.13.1",
+    "@gooddata/mock-js": "^2.14.2",
     "@gooddata/stylelint-config": "^5.0.0",
     "@gooddata/test-storybook": "4.0.0",
     "@gooddata/tslint-config": "1.0.0",
@@ -118,11 +118,11 @@
   },
   "dependencies": {
     "@formatjs/intl-pluralrules": "^1.3.6",
-    "@gooddata/gooddata-js": "^13.0.3",
+    "@gooddata/gooddata-js": "^13.0.4",
     "@gooddata/goodstrap": "^68.17.2",
     "@gooddata/js-utils": "^3.10.5",
     "@gooddata/numberjs": "^3.2.4",
-    "@gooddata/typings": "^2.25.0",
+    "@gooddata/typings": "^2.26.0",
     "@types/codemirror": "^0.0.95",
     "@types/mapbox-gl": "1.9.1",
     "ag-grid-community": "20.0.0",
