+++ src/internal/components/pluggableVisualizations/pivotTable/PluggablePivotTable.tsx
@@ -4,7 +4,7 @@ import get = require("lodash/get");
 import merge = require("lodash/merge");
 import flatMap = require("lodash/flatMap");
 import isNil = require("lodash/isNil");
-import includes = require("lodash/includes");
+import isEqual = require("lodash/isEqual");
 import * as React from "react";
 import ReactMeasure from "react-measure";
 import { render } from "react-dom";
@@ -18,7 +18,6 @@ import { unmountComponentsAtNodes } from "../../../utils/domHelper";
 import * as VisEvents from "../../../../interfaces/Events";
 import * as BucketNames from "../../../../constants/bucketNames";
 import {
-    IAttributeFilter,
     IBucket,
     IBucketFilter,
     IBucketItem,
@@ -26,7 +25,6 @@ import {
     IFeatureFlags,
     ILocale,
     IReferencePoint,
-    isAttributeFilter,
     IVisCallbacks,
     IVisConstruct,
     IVisProps,
@@ -52,6 +50,7 @@ import { AbstractPluggableVisualization } from "../AbstractPluggableVisualizatio
 import {
     getColumnWidthsFromProperties,
     getReferencePointWithSupportedProperties,
+    getSupportedProperties,
 } from "../../../utils/propertiesHelper";
 import { VisualizationEnvironment } from "../../../../components/uri/Visualization";
 import { VisualizationTypes } from "../../../../constants/visualizationTypes";
@@ -60,10 +59,18 @@ import { generateDimensions } from "../../../../helpers/dimensions";
 import { DEFAULT_LOCALE } from "../../../../constants/localization";
 import { DASHBOARDS_ENVIRONMENT } from "../../../constants/properties";
 import { ColumnWidthItem, IMenu, IPivotTableConfig } from "../../../../interfaces/PivotTable";
-import { adaptReferencePointWidthItemsToPivotTable } from "./widthItemsHelpers";
+import {
+    adaptReferencePointWidthItemsToPivotTable,
+    adaptMdObjectWidthItemsToPivotTable,
+} from "./widthItemsHelpers";
 import { PIVOT_TABLE_SUPPORTED_PROPERTIES } from "../../../constants/supportedProperties";
 
 import { getTableConfigFromFeatureFlags } from "../../../../helpers/featureFlags";
+import {
+    adaptMdObjectSortItemsToPivotTable,
+    adaptReferencePointSortItemsToPivotTable,
+    addDefaultSort,
+} from "./sortItemsHelpers";
 
 export const getColumnAttributes = (buckets: IBucket[]): IBucketItem[] => {
     return getItemsFromBuckets(
@@ -86,189 +93,6 @@ export const getRowAttributes = (buckets: IBucket[]): IBucketItem[] => {
         [ATTRIBUTE, DATE],
     );
 };
-
-// removes attribute sortItems with invalid identifiers
-// removes measure sortItems with invalid identifiers and invalid number of locators
-function adaptSortItemsToPivotTable(
-    originalSortItems: AFM.SortItem[],
-    measureLocalIdentifiers: string[],
-    rowAttributeLocalIdentifiers: string[],
-    columnAttributeLocalIdentifiers: string[],
-): AFM.SortItem[] {
-    const attributeLocalIdentifiers = [...rowAttributeLocalIdentifiers, ...columnAttributeLocalIdentifiers];
-
-    return originalSortItems.reduce((sortItems: AFM.SortItem[], sortItem: AFM.SortItem) => {
-        if (AFM.isMeasureSortItem(sortItem)) {
-            // filter out invalid locators
-            const filteredSortItem: AFM.IMeasureSortItem = {
-                measureSortItem: {
-                    ...sortItem.measureSortItem,
-                    locators: sortItem.measureSortItem.locators.filter(locator => {
-                        // filter out invalid measure locators
-                        if (AFM.isMeasureLocatorItem(locator)) {
-                            return includes(
-                                measureLocalIdentifiers,
-                                locator.measureLocatorItem.measureIdentifier,
-                            );
-                        }
-                        // filter out invalid column attribute locators
-                        return includes(
-                            columnAttributeLocalIdentifiers,
-                            locator.attributeLocatorItem.attributeIdentifier,
-                        );
-                    }),
-                },
-            };
-
-            // keep sortItem if measureLocator is present and locators are correct length
-            if (
-                filteredSortItem.measureSortItem.locators.some(locator =>
-                    AFM.isMeasureLocatorItem(locator),
-                ) &&
-                filteredSortItem.measureSortItem.locators.length ===
-                    columnAttributeLocalIdentifiers.length + 1
-            ) {
-                return [...sortItems, filteredSortItem];
-            }
-
-            // otherwise just carry over previous sortItems
-            return sortItems;
-        }
-        if (includes(attributeLocalIdentifiers, sortItem.attributeSortItem.attributeIdentifier)) {
-            return [...sortItems, sortItem];
-        }
-        return sortItems;
-    }, []);
-}
-
-export function adaptReferencePointSortItemsToPivotTable(
-    originalSortItems: AFM.SortItem[],
-    measures: IBucketItem[],
-    rowAttributes: IBucketItem[],
-    columnAttributes: IBucketItem[],
-): AFM.SortItem[] {
-    const measureLocalIdentifiers = measures.map(measure => measure.localIdentifier);
-    const rowAttributeLocalIdentifiers = rowAttributes.map(rowAttribute => rowAttribute.localIdentifier);
-    const columnAttributeLocalIdentifiers = columnAttributes.map(
-        columnAttribute => columnAttribute.localIdentifier,
-    );
-
-    return adaptSortItemsToPivotTable(
-        originalSortItems,
-        measureLocalIdentifiers,
-        rowAttributeLocalIdentifiers,
-        columnAttributeLocalIdentifiers,
-    );
-}
-
-const bucketItemGetter = <T extends VisualizationObject.BucketItem>(bucketId: string) => (
-    buckets: VisualizationObject.IBucket[],
-) => flatMap(buckets.filter(b => b.localIdentifier === bucketId), i => i.items) as T[];
-
-const getMeasures = bucketItemGetter<VisualizationObject.IMeasure>(BucketNames.MEASURES);
-const getRows = bucketItemGetter<VisualizationObject.IVisualizationAttribute>(BucketNames.ATTRIBUTE);
-const getColumns = bucketItemGetter<VisualizationObject.IVisualizationAttribute>(BucketNames.COLUMNS);
-
-function adaptMdObjectSortItemsToPivotTable(
-    originalSortItems: AFM.SortItem[],
-    buckets: VisualizationObject.IBucket[],
-): AFM.SortItem[] {
-    const measureLocalIdentifiers = getMeasures(buckets).map(measure => measure.measure.localIdentifier);
-    const rowAttributeLocalIdentifiers = getRows(buckets).map(
-        rowAttribute => rowAttribute.visualizationAttribute.localIdentifier,
-    );
-    const columnAttributeLocalIdentifiers = getColumns(buckets).map(
-        columnAttribute => columnAttribute.visualizationAttribute.localIdentifier,
-    );
-
-    return adaptSortItemsToPivotTable(
-        originalSortItems,
-        measureLocalIdentifiers,
-        rowAttributeLocalIdentifiers,
-        columnAttributeLocalIdentifiers,
-    );
-}
-
-const isAttributeSortItemVisible = (_sortItem: AFM.IAttributeSortItem, _filters: IBucketFilter[]): boolean =>
-    true;
-
-const isMeasureSortItemMatchedByFilter = (
-    sortItem: AFM.IMeasureSortItem,
-    filter: IAttributeFilter,
-): boolean =>
-    filter.selectedElements.some(selectedElement =>
-        sortItem.measureSortItem.locators.some(
-            locator =>
-                !AFM.isMeasureLocatorItem(locator) &&
-                locator.attributeLocatorItem.element === selectedElement.uri,
-        ),
-    );
-
-const isMeasureSortItemVisible = (sortItem: AFM.IMeasureSortItem, filters: IBucketFilter[]): boolean =>
-    filters.reduce((isVisible, filter) => {
-        if (isAttributeFilter(filter)) {
-            const shouldBeMatched = !filter.isInverted;
-            return isVisible && shouldBeMatched === isMeasureSortItemMatchedByFilter(sortItem, filter);
-        }
-        return isVisible;
-    }, true);
-
-export const isSortItemVisible = (sortItem: AFM.SortItem, filters: IBucketFilter[]): boolean =>
-    AFM.isAttributeSortItem(sortItem)
-        ? isAttributeSortItemVisible(sortItem, filters)
-        : isMeasureSortItemVisible(sortItem, filters);
-
-export function addDefaultSort(
-    sortItems: AFM.SortItem[],
-    filters: IBucketFilter[],
-    rowAttributes: IBucketItem[],
-    previousRowAttributes?: IBucketItem[],
-): AFM.SortItem[] {
-    // cannot construct default sort without a row
-    if (rowAttributes.length < 1) {
-        return sortItems;
-    }
-
-    // detect custom sort
-    const firstRow = rowAttributes[0];
-    const previousFirstRow = previousRowAttributes && previousRowAttributes[0];
-    const hasVisibleCustomSort = sortItems.some(sortItem => {
-        if (!isSortItemVisible(sortItem, filters)) {
-            return false;
-        }
-        // non attribute sort is definitely custom
-        if (!AFM.isAttributeSortItem(sortItem)) {
-            return true;
-        }
-        // asc sort on first row is considered default
-        if (
-            sortItem.attributeSortItem.attributeIdentifier === firstRow.localIdentifier &&
-            sortItem.attributeSortItem.direction === "asc"
-        ) {
-            return false;
-        }
-        // asc sort on row that was first until now is considered default as well
-        if (
-            previousFirstRow &&
-            sortItem.attributeSortItem.attributeIdentifier === previousFirstRow.localIdentifier &&
-            sortItem.attributeSortItem.direction === "asc"
-        ) {
-            return false;
-        }
-        return true;
-    });
-
-    return hasVisibleCustomSort
-        ? sortItems
-        : [
-              {
-                  attributeSortItem: {
-                      attributeIdentifier: firstRow.localIdentifier,
-                      direction: "asc",
-                  },
-              },
-          ];
-}
 
 export class PluggablePivotTable extends AbstractPluggableVisualization {
     private projectId: string;
@@ -309,9 +133,23 @@ export class PluggablePivotTable extends AbstractPluggableVisualization {
         mdObject: VisualizationObject.IVisualizationObjectContent,
         references: IReferences,
     ) {
-        this.visualizationProperties = visualizationProperties;
+        const propertiesWithOnlySupportedControls = getSupportedProperties(
+            visualizationProperties,
+            this.supportedPropertiesList,
+        );
+
+        const properties = visualizationProperties ? visualizationProperties.properties : {};
+
+        const onlySupportedProperties: IVisualizationPropertiesWrapper = {
+            properties: {
+                ...properties, // we are ignoring propertiesMeta
+                ...propertiesWithOnlySupportedControls,
+            },
+        };
+        this.visualizationProperties = onlySupportedProperties;
         this.references = references;
-        this.renderVisualization(options, visualizationProperties, mdObject);
+        this.adaptPropertiesToMdObject(onlySupportedProperties, mdObject);
+        this.renderVisualization(options, this.visualizationProperties, mdObject);
         this.renderConfigurationPanel(mdObject);
     }
 
@@ -599,6 +437,35 @@ export class PluggablePivotTable extends AbstractPluggableVisualization {
         return generateDimensions(mdObject, VisualizationTypes.TABLE);
     }
 
+    private adaptPropertiesToMdObject(
+        visualizationProperties: IVisualizationPropertiesWrapper,
+        mdObject: VisualizationObject.IVisualizationObjectContent,
+    ) {
+        // This is sanitization of properties from KD vs current mdObject from AD
+        const columnWidths = getColumnWidthsFromProperties(visualizationProperties);
+        if (columnWidths) {
+            this.sanitizeColumnWidths(columnWidths, mdObject);
+        }
+    }
+
+    private sanitizeColumnWidths(
+        columnWidths: ColumnWidthItem[],
+        mdObject: VisualizationObject.IVisualizationObjectContent,
+    ) {
+        const adaptedColumnWidths = adaptMdObjectWidthItemsToPivotTable(columnWidths, mdObject.buckets);
+        if (!isEqual(columnWidths, adaptedColumnWidths)) {
+            this.visualizationProperties.properties.controls.columnWidths = adaptedColumnWidths;
+            const { pushData } = this.callbacks;
+            pushData({
+                properties: {
+                    controls: {
+                        columnWidths: adaptedColumnWidths,
+                    },
+                },
+            });
+        }
+    }
+
     private enrichConfigWithMenu(config: IPivotTableConfig): IPivotTableConfig {
         if (this.environment === DASHBOARDS_ENVIRONMENT) {
             // Menu aggregations turned off in KD
@@ -612,31 +479,16 @@ export class PluggablePivotTable extends AbstractPluggableVisualization {
         return merge({ menu }, config);
     }
 
-    private getMergedProperties(newProperties: any): IVisualizationPropertiesWrapper {
-        const properties: IVisualizationPropertiesWrapper = get(
-            this.visualizationProperties,
-            "properties",
-            {},
-        ) as IVisualizationPropertiesWrapper;
-
-        return {
-            properties: {
-                ...properties,
-                ...newProperties,
-            },
-        };
-    }
-
     private onColumnResized(columnWidths: ColumnWidthItem[]) {
         const { pushData } = this.callbacks;
 
-        pushData(
-            this.getMergedProperties({
+        pushData({
+            properties: {
                 controls: {
                     columnWidths,
                 },
-            }),
-        );
+            },
+        });
     }
 
     private handlePushData(data: any) {
@@ -647,11 +499,11 @@ export class PluggablePivotTable extends AbstractPluggableVisualization {
         } = this;
         const properties = visualizationProperties ? visualizationProperties.properties : null;
         if (data && data.properties && data.properties.sortItems) {
-            pushData(
-                this.getMergedProperties({
+            pushData({
+                properties: {
                     sortItems: data.properties.sortItems,
-                }),
-            );
+                },
+            });
         } else if (data && data.result && properties && references) {
             // pushData from VisualizationLoadingHOC
             pushData({
