+++ src/internal/components/pluggableVisualizations/pivotTable/tests/PluggablePivotTable.spec.tsx
@@ -3,25 +3,15 @@ import * as React from "react";
 import * as ReactDom from "react-dom";
 import { AFM } from "@gooddata/typings";
 
-import {
-    adaptReferencePointSortItemsToPivotTable,
-    addDefaultSort,
-    getColumnAttributes,
-    getRowAttributes,
-    isSortItemVisible,
-    PluggablePivotTable,
-} from "../PluggablePivotTable";
+import { getColumnAttributes, getRowAttributes, PluggablePivotTable } from "../PluggablePivotTable";
 import * as testMocks from "../../../../mocks/testMocks";
 import * as referencePointMocks from "../../../../mocks/referencePointMocks";
 import * as uiConfigMocks from "../../../../mocks/uiConfigMocks";
 import {
     IBucket,
-    IBucketFilter,
-    IBucketFilterElement,
     IBucketItem,
     IExtendedReferencePoint,
     IFilters,
-    IFiltersBucketItem,
     ILocale,
     IVisConstruct,
     IVisProps,
@@ -45,208 +35,16 @@ import {
     validMeasureColumnWidthItem,
     transformedWeakMeasureColumnWidth,
 } from "./widthItemsMock";
-
-const getMockReferencePoint = (
-    measures: IBucketItem[] = [],
-    rows: IBucketItem[] = [],
-    columns: IBucketItem[] = [],
-    filterItems: IFiltersBucketItem[] = [],
-    sortItems: AFM.SortItem[] = [],
-    measuresIsShowInPercentEnabled = false,
-    columnWidths: ColumnWidthItem[] = [],
-): IExtendedReferencePoint => ({
-    buckets: [
-        {
-            items: measures,
-            localIdentifier: "measures",
-        },
-        {
-            items: rows,
-            localIdentifier: "attribute",
-        },
-        {
-            items: columns,
-            localIdentifier: "columns",
-        },
-    ],
-    filters: {
-        items: filterItems,
-        localIdentifier: "filters",
-    },
-    properties: {
-        sortItems,
-        controls: {
-            columnWidths,
-        },
-    },
-    uiConfig: {
-        buckets: {
-            attribute: {
-                accepts: ["attribute", "date"],
-                allowsReordering: true,
-                allowsSwapping: true,
-                canAddItems: true,
-                enabled: true,
-                icon: "",
-                isShowInPercentEnabled: false,
-                itemsLimit: 20,
-                title: "Rows",
-            },
-            columns: {
-                accepts: ["attribute", "date"],
-                allowsReordering: true,
-                allowsSwapping: true,
-                canAddItems: true,
-                enabled: true,
-                icon: "",
-                isShowInPercentEnabled: false,
-                itemsLimit: 20,
-                title: "Columns",
-            },
-            filters: {
-                accepts: ["attribute", "date"],
-                allowsReordering: false,
-                enabled: true,
-                isShowInPercentEnabled: false,
-                itemsLimit: 20,
-            },
-            measures: {
-                accepts: ["metric", "fact", "attribute"],
-                allowsDuplicateItems: true,
-                allowsReordering: true,
-                allowsSwapping: true,
-                canAddItems: true,
-                enabled: true,
-                icon: "",
-                isShowInPercentEnabled: measuresIsShowInPercentEnabled,
-                isShowInPercentVisible: true,
-                itemsLimit: 20,
-                title: "Measures",
-            },
-        },
-        exportConfig: {
-            supported: true,
-        },
-        noMetricAccepted: {
-            supported: true,
-        },
-        openAsReport: {
-            supported: false,
-        },
-        recommendations: {},
-        supportedOverTimeComparisonTypes: ["same_period_previous_year", "previous_period"],
-    },
-});
-
-// sortItems based on referencePointMocks.simpleStackedReferencePoint
-const validMeasureSort: AFM.IMeasureSortItem = {
-    measureSortItem: {
-        direction: "asc",
-        locators: [
-            {
-                attributeLocatorItem: {
-                    attributeIdentifier: "a2",
-                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
-                },
-            },
-            {
-                measureLocatorItem: {
-                    measureIdentifier: "m1",
-                },
-            },
-        ],
-    },
-};
-
-const validAttributeSort: AFM.IAttributeSortItem = {
-    attributeSortItem: {
-        attributeIdentifier: "a1",
-        direction: "desc",
-    },
-};
-
-const invalidAttributeSort: AFM.IAttributeSortItem = {
-    attributeSortItem: {
-        attributeIdentifier: "invalid",
-        direction: "desc",
-    },
-};
-
-const invalidMeasureSortInvalidMeasure: AFM.IMeasureSortItem = {
-    measureSortItem: {
-        direction: "asc",
-        locators: [
-            {
-                attributeLocatorItem: {
-                    attributeIdentifier: "a2",
-                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
-                },
-            },
-            {
-                measureLocatorItem: {
-                    measureIdentifier: "invalid", // this measure is not in buckets
-                },
-            },
-        ],
-    },
-};
-
-const invalidMeasureSortInvalidAttribute: AFM.IMeasureSortItem = {
-    measureSortItem: {
-        direction: "asc",
-        locators: [
-            {
-                attributeLocatorItem: {
-                    attributeIdentifier: "a1", // this identifier doesn't exist on second dimension
-                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
-                },
-            },
-            {
-                measureLocatorItem: {
-                    measureIdentifier: "m1",
-                },
-            },
-        ],
-    },
-};
-
-const invalidMeasureSortTooManyLocators: AFM.IMeasureSortItem = {
-    measureSortItem: {
-        direction: "asc",
-        locators: [
-            {
-                attributeLocatorItem: {
-                    attributeIdentifier: "a2",
-                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
-                },
-            },
-            {
-                attributeLocatorItem: {
-                    attributeIdentifier: "a2",
-                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
-                },
-            },
-            {
-                measureLocatorItem: {
-                    measureIdentifier: "m1",
-                },
-            },
-        ],
-    },
-};
-
-const invalidMeasureSortLocatorsTooShort: AFM.IMeasureSortItem = {
-    measureSortItem: {
-        direction: "asc",
-        locators: [
-            {
-                measureLocatorItem: {
-                    measureIdentifier: "m1",
-                },
-            },
-        ],
-    },
-};
+import {
+    invalidAttributeSort,
+    invalidMeasureSortInvalidMeasure,
+    invalidMeasureSortInvalidAttribute,
+    invalidMeasureSortLocatorsTooShort,
+    invalidMeasureSortTooManyLocators,
+    validAttributeSort,
+    validMeasureSort,
+} from "./sortItemsMock";
+import { getMockReferencePoint } from "./testHelper";
 
 describe("PluggablePivotTable", () => {
     const defaultProps = {
@@ -461,7 +259,7 @@ describe("PluggablePivotTable", () => {
                 {
                     attributeColumnWidthItem: {
                         width: 740,
-                        attributeIdentifier: "294512a6b2ed4be8bd3948dd14db1950",
+                        attributeIdentifier: "a1",
                     },
                 },
             ];
@@ -515,7 +313,12 @@ describe("PluggablePivotTable", () => {
                     };
 
                     const options = getDefaultOptions();
-                    pivotTable.update({ ...options }, visualizationProperties, testMocks.emptyMdObject, null);
+                    pivotTable.update(
+                        { ...options },
+                        visualizationProperties,
+                        testMocks.mdObjectAttributeOnly,
+                        null,
+                    );
 
                     const props: any = createTableSpy.mock.calls[0][0];
 
@@ -1010,311 +813,5 @@ describe("getRowAttributes", () => {
                 type: "attribute",
             },
         ]);
-    });
-});
-
-describe("adaptReferencePointSortItemsToPivotTable", () => {
-    const sourceReferencePoint = referencePointMocks.simpleStackedReferencePoint;
-    const mockPivotTableReferencePoint: IExtendedReferencePoint = getMockReferencePoint(
-        sourceReferencePoint.buckets[0].items,
-        sourceReferencePoint.buckets[1].items,
-        sourceReferencePoint.buckets[2].items,
-        [],
-        [],
-        true,
-    );
-
-    const sourceSortItems: AFM.SortItem[] = [
-        invalidAttributeSort,
-        invalidMeasureSortInvalidMeasure,
-        invalidMeasureSortInvalidAttribute,
-        invalidMeasureSortLocatorsTooShort,
-        invalidMeasureSortTooManyLocators,
-        validAttributeSort,
-        validMeasureSort,
-    ];
-
-    const measures: IBucketItem[] = mockPivotTableReferencePoint.buckets[0].items;
-    const rowAttributes: IBucketItem[] = mockPivotTableReferencePoint.buckets[1].items;
-    const columnAttributes: IBucketItem[] = mockPivotTableReferencePoint.buckets[2].items;
-
-    it("should remove invalid sort items", async () => {
-        const expectedSortItems: AFM.SortItem[] = [validAttributeSort, validMeasureSort];
-
-        expect(
-            adaptReferencePointSortItemsToPivotTable(
-                sourceSortItems,
-                measures,
-                rowAttributes,
-                columnAttributes,
-            ),
-        ).toEqual(expectedSortItems);
-    });
-});
-
-describe("addDefaultSort", () => {
-    const rowFor = (localIdentifier: string, attributeName: string): IBucketItem => ({
-        aggregation: null,
-        showInPercent: null,
-        operator: null,
-        operandLocalIdentifiers: null,
-        granularity: null,
-        masterLocalIdentifier: null,
-        localIdentifier,
-        showOnSecondaryAxis: null,
-        type: "attribute",
-        filters: [],
-        attribute: attributeName,
-    });
-
-    const accountRowId = "d0f1043a2eec42daac004a10c41afdd5";
-    const accountRow = rowFor(accountRowId, "attr.account");
-
-    const countryRowId = "6d523113ca754409a8b685736e7fe32b";
-    const countryRow = rowFor(countryRowId, "attr.country");
-
-    const productRowId = "38f1d87b8b7c42c1b9a37395a24f7313";
-    const productRow = rowFor(productRowId, "attr.product");
-
-    const sortFor = (localId: string, direction: AFM.SortDirection): AFM.SortItem => ({
-        attributeSortItem: {
-            attributeIdentifier: localId,
-            direction,
-        },
-    });
-
-    const defaultSortFor = (localId: string): AFM.SortItem => sortFor(localId, "asc");
-
-    describe("with no filters specified", () => {
-        it("should not add the default sort if no row is specified", () => {
-            const expected: AFM.SortItem[] = [];
-            const actual = addDefaultSort([], [], [], []);
-            expect(actual).toEqual(expected);
-        });
-        it("should add the default sort for the first row added", () => {
-            const expected = [defaultSortFor(accountRowId)];
-            const actual = addDefaultSort([], [], [accountRow], []);
-            expect(actual).toEqual(expected);
-        });
-        it("should add the default sort when a row is added to the first place", () => {
-            const expected = [defaultSortFor(accountRowId)];
-            const actual = addDefaultSort(
-                [defaultSortFor(countryRowId)],
-                [],
-                [accountRow, countryRow],
-                [countryRow],
-            );
-            expect(actual).toEqual(expected);
-        });
-        it("should not change the default sort when a row is added to the second place", () => {
-            const expected = [defaultSortFor(countryRowId)];
-            const actual = addDefaultSort(
-                [defaultSortFor(countryRowId)],
-                [],
-                [countryRow, accountRow],
-                [countryRow],
-            );
-            expect(actual).toEqual(expected);
-        });
-        it("should not change sorts when there is a desc sort on the first item", () => {
-            const expected = [sortFor(countryRowId, "desc")];
-            const actual = addDefaultSort(
-                [sortFor(countryRowId, "desc")],
-                [],
-                [countryRow, accountRow],
-                [countryRow],
-            );
-            expect(actual).toEqual(expected);
-        });
-        it("should not change sorts when there is a desc sort on the second item", () => {
-            const expected = [sortFor(countryRowId, "desc")];
-            const actual = addDefaultSort(
-                [sortFor(countryRowId, "desc")],
-                [],
-                [accountRow, countryRow, productRow],
-                [accountRow, countryRow],
-            );
-            expect(actual).toEqual(expected);
-        });
-        it("should not change sorts when there is an asc sort on the second item", () => {
-            const expected = [sortFor(countryRowId, "asc")];
-            const actual = addDefaultSort(
-                [sortFor(countryRowId, "asc")],
-                [],
-                [accountRow, countryRow, productRow],
-                [accountRow, countryRow],
-            );
-            expect(actual).toEqual(expected);
-        });
-        it("should not change sorts when there is a measure sort", () => {
-            const measureSort: AFM.IMeasureSortItem = {
-                measureSortItem: {
-                    direction: "asc",
-                    locators: [],
-                },
-            };
-            const expected = [measureSort];
-            const actual = addDefaultSort(
-                [measureSort],
-                [],
-                [accountRow, countryRow, productRow],
-                [accountRow, countryRow],
-            );
-            expect(actual).toEqual(expected);
-        });
-    });
-
-    it("should add default sort if existing measure sort is not visible due to filters (RAIL-1275)", () => {
-        const expected = [sortFor(accountRowId, "asc")];
-        const uri = "/gdc/md/mockproject/obj/attr.country/elements?id=1";
-        const measureSort: AFM.IMeasureSortItem = {
-            measureSortItem: {
-                direction: "asc",
-                locators: [
-                    {
-                        attributeLocatorItem: {
-                            attributeIdentifier: "attr.country",
-                            element: uri,
-                        },
-                    },
-                ],
-            },
-        };
-        const filterElement: IBucketFilterElement = {
-            title: "Matching",
-            uri,
-        };
-        const actual = addDefaultSort(
-            [measureSort],
-            [
-                {
-                    attribute: "irrelevant",
-                    isInverted: true,
-                    selectedElements: [filterElement],
-                    totalElementsCount: 4,
-                },
-            ],
-            [accountRow, countryRow, productRow],
-            [accountRow, countryRow],
-        );
-        expect(actual).toEqual(expected);
-    });
-});
-
-describe("isSortItemVisible", () => {
-    describe("given attribute sort item", () => {
-        it("should always return true", () => {
-            const actual = isSortItemVisible(
-                {
-                    attributeSortItem: {
-                        attributeIdentifier: "foo",
-                        direction: "asc",
-                    },
-                },
-                [],
-            );
-            const expected = true;
-            expect(actual).toEqual(expected);
-        });
-    });
-
-    describe("given measure sort item", () => {
-        const createFilterBucketItem = (
-            selectedElements: IBucketFilterElement[],
-            isInverted: boolean,
-        ): IBucketFilter => ({
-            attribute: "irrelevant",
-            isInverted,
-            totalElementsCount: 5,
-            selectedElements,
-        });
-
-        const matchingUri = "/gdc/md/mockproject/obj/attr.movie_genre/elements?id=1";
-        const notMatchingUri = "/gdc/md/mockproject/obj/attr.movie_genre/elements?id=123";
-        const sortItem: AFM.IMeasureSortItem = {
-            measureSortItem: {
-                direction: "asc",
-                locators: [
-                    {
-                        attributeLocatorItem: {
-                            attributeIdentifier: "foo",
-                            element: matchingUri,
-                        },
-                    },
-                    {
-                        measureLocatorItem: {
-                            measureIdentifier: "bar",
-                        },
-                    },
-                ],
-            },
-        };
-        const matchingFilterElement: IBucketFilterElement = {
-            title: "Matching",
-            uri: matchingUri,
-        };
-        const notMatchingFilterElement: IBucketFilterElement = {
-            title: "Not Matching",
-            uri: notMatchingUri,
-        };
-        const measureValueFilter: IBucketFilter = {
-            measureLocalIdentifier: "id",
-            condition: {
-                range: {
-                    operator: "BETWEEN",
-                    from: 0,
-                    to: 0,
-                },
-            },
-        };
-
-        it("should return true when no filters are specified", () => {
-            const actual = isSortItemVisible(sortItem, []);
-            const expected = true;
-            expect(actual).toEqual(expected);
-        });
-        it('should return true when empty "notIn" filter is specified', () => {
-            const actual = isSortItemVisible(sortItem, [createFilterBucketItem([], true)]);
-            const expected = true;
-            expect(actual).toEqual(expected);
-        });
-        it('should return false when "notIn" filter with matching element is specified', () => {
-            const actual = isSortItemVisible(sortItem, [
-                createFilterBucketItem([matchingFilterElement], true),
-            ]);
-            const expected = false;
-            expect(actual).toEqual(expected);
-        });
-        it('should return true when "notIn" filter without matching element is specified', () => {
-            const actual = isSortItemVisible(sortItem, [
-                createFilterBucketItem([notMatchingFilterElement], true),
-            ]);
-            const expected = true;
-            expect(actual).toEqual(expected);
-        });
-
-        it('should return false when empty "in" filter is specified', () => {
-            const actual = isSortItemVisible(sortItem, [createFilterBucketItem([], false)]);
-            const expected = false;
-            expect(actual).toEqual(expected);
-        });
-        it('should return true when "in" filter with matching element is specified', () => {
-            const actual = isSortItemVisible(sortItem, [
-                createFilterBucketItem([matchingFilterElement], false),
-            ]);
-            const expected = true;
-            expect(actual).toEqual(expected);
-        });
-        it('should return false when "in" filter without matching element is specified', () => {
-            const actual = isSortItemVisible(sortItem, [
-                createFilterBucketItem([notMatchingFilterElement], false),
-            ]);
-            const expected = false;
-            expect(actual).toEqual(expected);
-        });
-        it("should return true when filter is MVF", () => {
-            const actual = isSortItemVisible(sortItem, [measureValueFilter]);
-            expect(actual).toEqual(true);
-        });
     });
 });
